Move simulation over to golfsimulator.py
